FAI-3971: Batch mutations in Faros Destination 1) review comments
